Inline GLM KDA reference in its parity test
